Move command files
