+++ canvas/cli/__init__.py
@@ -2,4 +2,7 @@ from __future__ import annotations
 
 # flake8: noqa
 from .base import *
+from .init import *
+from .add import *
+from .status import *
 from .manager import *
