+++ canvas/cli/add.py
