--- canvas/cli/commands/__init__.py
@@ -1,6 +1,0 @@
-from __future__ import annotations
-
-# flake8: noqa
-from .init import *
-from .add import *
-from .status import *
